+++ app/src/main/java/com/github/orkest/Constants.kt
@@ -29,5 +29,8 @@ class Constants {
         val DUMMY_LAST_CONNECTED_TIME: LocalDateTime =
             LocalDateTime.of(2021, 5, 1, 12, 0)
 
+        val NOTIFICATION_CHANNEL_ID = "your_channel_id"
+
+
     }
 }
